backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & Auth
supabase==2.3.0
postgrest==0.13.2

# AI & LLM
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
llama-index==0.9.48

# Vector DB (optional)
# pinecone-client==3.0.0
# weaviate-client==4.4.0

# HTTP & API
httpx==0.26.0
requests==2.31.0
python-multipart==0.0.6

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
